% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\name{clone_env}
\alias{clone_env}
\title{Clones an environment, either a deep or shallow copy.}
\usage{
clone_env(env, deep = FALSE)
}
\arguments{
\item{env}{The environment to clone}

\item{deep}{Whether to perform a deep copy}
}
\value{
The cloned environment
}
\description{
Clones an environment, either a deep or shallow copy.
}
