% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_collector.R
\name{collector_env}
\alias{collector_env}
\title{An environment into which a collector will be initialized (so we don't need to update
global variables).}
\usage{
collector_env
}
\description{
An environment into which a collector will be initialized (so we don't need to update
global variables).
}
