% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_collector.R
\name{collector_varname}
\alias{collector_varname}
\title{The name of the active collector variable in \link{collector_env}}
\usage{
collector_varname
}
\description{
The name of the active collector variable in \link{collector_env}
}
