% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export a JupyterNotebook to a zip file}
\usage{
export(notebook_path, export_path = NULL, display_link = TRUE)
}
\arguments{
\item{notebook_path}{The path to the notebook}

\item{export_path}{The path at which to write the zip file (optional)}

\item{display_link}{Whether to display a download link with \code{IRdisplay}}
}
\description{
Export a Jupyter Notebook to a zip file for submission.
}
\examples{
\dontrun{
export("hw01.ipynb")
}
}
