% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{grade_script}
\alias{grade_script}
\title{Execute a script, parse check outputs, and run additional tests specified by the glob pattern
\code{tests_glob} on the test environment.}
\usage{
grade_script(script_path, tests_glob, secret)
}
\arguments{
\item{script_path}{The path to the script}

\item{tests_glob}{The pattern to search for extra tests}

\item{secret}{The string to be appended to the name \code{check_results_} as the list name to collect
results (optional)}
}
\value{
The list of test_suite_result objects after executing tests referenced in the script
and those specified by \code{tests_glob}
}
\description{
Execute a script, parse check outputs, and run additional tests specified by the glob pattern
\code{tests_glob} on the test environment.
}
