% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_cases.R
\name{load_test_cases}
\alias{load_test_cases}
\title{Load test cases from a test file}
\usage{
load_test_cases(test_file)
}
\arguments{
\item{test_file}{The path to the test file}
}
\value{
The test cases
}
\description{
Load test case data from a test file. Executes the file and grabs the global \code{test}
variable, which should be a \code{list}.
}
