% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullish}
\alias{nullish}
\title{Check whether a value is nullish (i.e. NULL or NA)}
\usage{
nullish(x)
}
\arguments{
\item{x}{The value to cehck}
}
\value{
Whether the value is nullish
}
\description{
Check whether a value is nullish (i.e. NULL or NA)
}
