% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{parse_suite_results}
\alias{parse_suite_results}
\title{Parse output from \code{testthat::ListReporter} and return an instance of the ref class test_suite_results
constructed from this output}
\usage{
parse_suite_results(suite_results, test_metadata, num_cases)
}
\arguments{
\item{suite_results}{The output from a \code{testthat::ListReporter} as a list}

\item{test_metadata}{The parsed metadata from the test suite}

\item{num_cases}{The number of test cases in the suite}
}
\value{
The parsed results for the test suite
}
\description{
Parse output from \code{testthat::ListReporter} and return an instance of the ref class test_suite_results
constructed from this output
}
