% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{running_on_jupyter}
\alias{running_on_jupyter}
\title{Determine whether this R session is running on Jupyter.}
\usage{
running_on_jupyter()
}
\value{
A boolean indicating whether IRkernel is running.
}
\description{
Determine whether this R session is running on Jupyter by checking for a CommManager
in IRkernel.
}
\examples{
\dontrun{
running_on_jupyter()
}
}
