% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_notebook.R
\name{save_notebook}
\alias{save_notebook}
\title{Attempt to save the current notebook.}
\usage{
save_notebook(nb_path, timeout = 10)
}
\arguments{
\item{nb_path}{The path to the notebook}

\item{timeout}{Number of seconds to wait for save}
}
\value{
A boolean indicating whether the file was saved successfully. If
Jupyter is not running, this function returns TRUE.
}
\description{
Attempt to save the notebook by displaying Javascript if running on Jupyter. This
function waits until the modification time of the file has changed or until the specified timeout
expires.
}
\examples{
\dontrun{
save_notebook("foo.ipynb")
}
}
