% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\docType{data}
\name{VALID_EXPR_CHARS}
\alias{VALID_EXPR_CHARS}
\title{A string containing characters that can be made into a valid variable name. Does not include any
digits because randomly sampling with them included could result in an invalid variable name.}
\format{
An object of class \code{character} of length 1.
}
\usage{
VALID_EXPR_CHARS
}
\description{
A string containing characters that can be made into a valid variable name. Does not include any
digits because randomly sampling with them included could result in an invalid variable name.
}
\keyword{datasets}
