% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_collector.R
\docType{data}
\name{collector_env}
\alias{collector_env}
\title{An environment into which a collector will be initialized (so we don't need to update
global variables).}
\format{
An object of class \code{environment} of length 0.
}
\usage{
collector_env
}
\description{
An environment into which a collector will be initialized (so we don't need to update
global variables).
}
\keyword{datasets}
