% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_collector.R
\docType{data}
\name{collector_varname}
\alias{collector_varname}
\title{The name of the active collector variable in \link{collector_env}}
\format{
An object of class \code{character} of length 1.
}
\usage{
collector_varname
}
\description{
The name of the active collector variable in \link{collector_env}
}
\keyword{datasets}
