% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export a submission to a zip file}
\usage{
export(
  submission_path,
  export_path = NULL,
  display_link = TRUE,
  pdf = FALSE,
  force_save = FALSE
)
}
\arguments{
\item{submission_path}{The path to the submission}

\item{export_path}{The path at which to write the zip file (optional)}

\item{display_link}{Whether to display a download link with \code{IRdisplay}}

\item{pdf}{Whether to include a PDF of the submission (only works for Rmd and ipynb files)}

\item{force_save}{Whether to attempt to force-save the notebook if running on Jupyter}
}
\description{
Export a submission to a zip file for submitting. If indicated, a PDF of the
submission is generated and included in the zip file. (PDF generation is only supported for Rmd
and ipynb files.)
}
\examples{
\dontrun{
export("hw01.ipynb")

# with pdf
export("hw01.ipynb", pdf = TRUE)
}
}
