% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_script.R
\name{grade_script}
\alias{grade_script}
\title{Grade an R script against a series of test files}
\usage{
grade_script(script_path, tests_glob, ignore_errors)
}
\arguments{
\item{script_path}{The path to the script}

\item{tests_glob}{The pattern to search for extra tests}

\item{ignore_errors}{Whether to ignore errors thrown while executing the script}
}
\value{
The \link{GradingResults} object after executing tests referenced in the script
and those specified by \code{tests_glob}
}
\description{
Execute a script, parse check outputs, and run additional tests specified by the
glob pattern \code{tests_glob} on the test environment.
}
