% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\name{load_test_cases}
\alias{load_test_cases}
\title{Loads test case data from a test file. Executes the file and grabs the global \code{test} variable,
which should be a \code{list}.}
\usage{
load_test_cases(test_file)
}
\arguments{
\item{test_file}{The path to the test file}
}
\value{
The test cases
}
\description{
Loads test case data from a test file. Executes the file and grabs the global \code{test} variable,
which should be a \code{list}.
}
