% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\name{results_to_json}
\alias{results_to_json}
\title{Export a list of \code{TestFileResult} objects to a JSON string}
\usage{
results_to_json(results)
}
\arguments{
\item{results}{The list of result objects}
}
\value{
The JSON string
}
\description{
Export a list of \code{TestFileResult} objects to a JSON string
}
