% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter.r
\name{results_to_list}
\alias{results_to_list}
\title{Convert a list of \code{test_suite_result} objects to a JSON-like object of the correct form for writing
results for Gradescope.}
\usage{
results_to_list(results)
}
\arguments{
\item{results}{The list of \code{test_suite_result}s}
}
\value{
The generated list
}
\description{
Convert a list of \code{test_suite_result} objects to a JSON-like object of the correct form for writing
results for Gradescope.
}
