% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_autograder.R
\name{run_autograder}
\alias{run_autograder}
\title{Grade an R script against test files in a directory}
\usage{
run_autograder(script_path, secret, ignore_errors, test_dir)
}
\arguments{
\item{script_path}{The path to the script}

\item{secret}{The string to be appended to the name \code{check_results_} as the list name to collect
results (optional)}

\item{ignore_errors}{Whether to ignore errors thrown while executing the script}

\item{test_dir}{A directory of tests to glob from}
}
\value{
The JSON string
}
\description{
Run autograder in a Gradescope container and return the results as a
properly-formatted JSON string.
}
\examples{
\dontrun{
run_autograder("hw01.R", "ABC123", TRUE, "tests")
}
}
