% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\name{TestCaseResult}
\alias{TestCaseResult}
\title{A utility class for tracking the results of a \code{TestCase}}
\description{
A utility class for tracking the results of a \code{TestCase}

A utility class for tracking the results of a \code{TestCase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TestCaseResult$new()}}
\item \href{#method-get_score}{\code{TestCaseResult$get_score()}}
\item \href{#method-repr}{\code{TestCaseResult$repr()}}
\item \href{#method-to_list}{\code{TestCaseResult$to_list()}}
\item \href{#method-get_message}{\code{TestCaseResult$get_message()}}
\item \href{#method-clone}{\code{TestCaseResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseResult$new(passed, error, test_case)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{passed}}{Whether the test passed}

\item{\code{error}}{An error raised by executing the test, if any}

\item{\code{test_case}}{The \code{TestCase} that this result tracks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_score"></a>}}
\if{latex}{\out{\hypertarget{method-get_score}{}}}
\subsection{Method \code{get_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseResult$get_score()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseResult$repr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseResult$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_message"></a>}}
\if{latex}{\out{\hypertarget{method-get_message}{}}}
\subsection{Method \code{get_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseResult$get_message()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestCaseResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
