% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\name{TestFileResult}
\alias{TestFileResult}
\title{A collection of test case results that correspond to a single test file}
\description{
A collection of test case results that correspond to a single test file

A collection of test case results that correspond to a single test file
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TestFileResult-new}{\code{TestFileResult$new()}}
\item \href{#method-TestFileResult-get_basename}{\code{TestFileResult$get_basename()}}
\item \href{#method-TestFileResult-get_score}{\code{TestFileResult$get_score()}}
\item \href{#method-TestFileResult-repr}{\code{TestFileResult$repr()}}
\item \href{#method-TestFileResult-failed_hidden_cases}{\code{TestFileResult$failed_hidden_cases()}}
\item \href{#method-TestFileResult-failed_public_cases}{\code{TestFileResult$failed_public_cases()}}
\item \href{#method-TestFileResult-get_points}{\code{TestFileResult$get_points()}}
\item \href{#method-TestFileResult-failed_any_public}{\code{TestFileResult$failed_any_public()}}
\item \href{#method-TestFileResult-to_list}{\code{TestFileResult$to_list()}}
\item \href{#method-TestFileResult-clone}{\code{TestFileResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$new(test_case_results, filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_case_results}}{The \code{TestCaseResult} objects that make up this test file}

\item{\code{filename}}{The name of the test file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-get_basename"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-get_basename}{}}}
\subsection{Method \code{get_basename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$get_basename()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-get_score"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-get_score}{}}}
\subsection{Method \code{get_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$get_score()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-repr"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-repr}{}}}
\subsection{Method \code{repr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$repr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-failed_hidden_cases"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-failed_hidden_cases}{}}}
\subsection{Method \code{failed_hidden_cases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$failed_hidden_cases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-failed_public_cases"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-failed_public_cases}{}}}
\subsection{Method \code{failed_public_cases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$failed_public_cases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-get_points"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-get_points}{}}}
\subsection{Method \code{get_points()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$get_points()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-failed_any_public"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-failed_any_public}{}}}
\subsection{Method \code{failed_any_public()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$failed_any_public()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFileResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TestFileResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFileResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
