% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottr.r
\name{valid_syntax}
\alias{valid_syntax}
\title{Determine whether a code snippet has any syntax errors.}
\usage{
valid_syntax(script)
}
\arguments{
\item{script}{The code snippet}
}
\value{
Whether the code snippet is valid (can be parsed with \code{parse})
}
\description{
Determine whether a code snippet has any syntax errors.
}
